requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Jinja2==3.0.3
Werkzeug==2.0.3
itsdangerous==2.0.1
click==8.0.4

SQLAlchemy==1.4.53
Alembic==1.12.0
psycopg2==2.9.7
psycopg2-binary==2.9.9

python-dateutil==2.8.2
pytz==2023.3.post1
python-dotenv==1.0.1
PyYAML==6.0.1
openpyxl==3.1.5
Flasgger==0.9.7.1
marshmallow==3.20.1
Unidecode==1.3.8
requests==2.31.0

pytest==7.4.0
